refactor: moved utils and types into src instead of their own folder. removed the main index.ts since nothing from there is being exported anymore. made a seperate plugin folder for the matrixai eslint plugin that contains our custom rules
